Reimplement with class and recursivity
For keeping an state (the root dictionary) and being a tree a recursive structure by nature, and in this case of words, not so deep structure, I adventured in the reimplementation with:

 * A Calss (for state)
 * Recursivity (for tree)

The real value is perceived by the client code. See the last 4 lines of code. Now you obtain a Trie with `Trie(words)` instead of `generateTrieFromWordsArray(arrayOfWords)` and you can test if a word is in the Trie with `'hello' in trie` instead of `isWordPresentInTrie(wordsTrie, 'hello')`.

Also, I removed the special character for end of tree. An empty string when analyzing the word should be considered as the word is in the trie. This imply that the empty string "is in" every trie. I prefered assume this rather than code a especial case. Should this assumption be wrong, the patch should go in the `__contains__` method.